requirements.txt
tabula-py>=2.9.0
pdfplumber>=0.11.0
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
tqdm>=4.66.0
click>=8.1.0